added roles. now the backend can differentiate between developers and normal users. Normal users: can browse games, buy games.
Developers: can upload, update, or delete their own games.

Without this role in the token, the server wouldn’t know if a logged-in person should be allowed to upload a game or not.

we introduced role in const userSchema in User.js and also modified made changes in signup and login in auth.js